requirements.txt
agentai==0.0.4
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
docstring-parser==0.15
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
geneimpacts==0.3.7
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.16.3
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
huggingface-hub==0.23.0
idna==3.7
inheritance==0.1.5
ipython==8.24.0
ipython-genutils==0.2.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
netifaces==0.11.0
numpy==1.26.4
<<<<<<< Updated upstream
openpyxl==2.4.2
=======
openai==0.27.10
openpyxl==1.8.6
>>>>>>> Stashed changes
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic==1.10.15
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
<<<<<<< Updated upstream
requests==2.32.2
=======
regex==2024.5.10
requests==2.31.0
>>>>>>> Stashed changes
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.0.272
simplegeneric==0.8.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.34.0
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
yarl==1.9.4
